terraform/01-bedrock-agents/scripts/requirements.txt
boto3>=1.42.0
botocore[crt]>=1.42.0
python-dotenv>=1.0.0
